Code/RATIO_MCP/requirements.txt
python-dotenv
azure-identity
autogen-ext
pyodbc
azure-kusto-data
azure-core>=1.0.0
autogen-agentchat
dotenv
openai
tiktoken
autogen-ext[openai,azure,mcp,docker]
streamlit>=1.38.0
rapidfuzz>=3.0.0
msal
requests
pandas
altair
plotly
PyYAML
azure-keyvault-secrets
fastapi
uvicorn
pydantic
azure-keyvault
mcp>=1.5.0
azure-identity>=1.17.0
azure-storage-file-datalake>=12.16.0
PyJWT>=2.9.0
fastmcp
azure-monitor-opentelemetry